ci-eks: Add IPsec key rotation tests
[ upstream commit 5c06c8e2ea5c575276f84d7f90d967f3c9cb3551 ]

First, this commit includes the IPsec key rotation tests action.

Second, it changes the CLI exec name and path to "./cilium-cli", so that
it can be used by the key rotation action and friends.

Third, it runs the IPsec tests only if the matrix.ipsec is set to
"true". A subsequent commit will extend the matrix configuration
accordingly.

Signed-off-by: Martynas Pumputis <m@lambda.lt>
Signed-off-by: Julian Wiedmann <jwi@isovalent.com>