azure/ipam: fix bug where vm ip can be allocated twice due to race.
TestIpamManyNodes fails occasionally with unexpected available IP counts.
This appears to be an underlying bug in the azure/ipam implementation.
This results from two racing resyncInstance calls racing to update the
cached instances list.

The chain of triggers for Upsert is as follows:

*note*: Node has a reference to nodemanager, and nodemanager maintains
	references to all nodes, thus each node has a circular ref w/
	the node manager.  This complicates locking.

nodemanager.Upsert ->
	1. (NodeManager) Sets up node triggers.
	2. (Node) Does node IP IPAM accounting.
	3. (Node) Triggers poolMaintenance.
poolMaintenance ->
	1. (Node) Attach any necessary private IPs to instance based on current
		accounting (i.e. node stats).
	2. (Node) Performs node device/address accounting and triggers instance
		sync (i.e. in this case: syncing Cloud API instance/device/addr with
		local cache).
instanceSync ->
	1. (NodeManager/InstanceManager) API is called to get
		subnets and instances.
	2. (NodeManager/InstanceManager) Instances are updated in
	   	critical section.

Finally, based on the outcome of the instances sync, following addr
accounting, another maintainIPPool may be triggered causing a series of
confusing allocation/accounting errors.

In this last part, because getting instance data from the API is not
locked, and because the instanceSync operation is an operation on the
NodeManager. Two such syncs can occur at the same time (i.e. from
different vms), in such a case, two instanceSync runs may occur out of
order and the most recent result will be overwritten.

This moves the azure API implementation lock to encompass the entire
operation Resync operation - such that getting instances/subnets and
assigning them is atomic.

This shouldn't have any risk of deadlocks as azure api doesn't have any
references to node or node manager.

Fixes: #31466

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>