ci/ipsec: Print more info to debug credentials removal check failures
[ upstream commit 129f2e235e62445b73a1b5630f1f7a3a36bf5014 ]

In commit 6fee46f9e753 ("ci/ipsec: Fix downgrade version retrieval") we
added a check to make sure that GitHub credentials are removed before
pulling the untrusted branch from the Pull Request's author. It appears
that this check occasionally fails and causes the whole job to abort.
But Cilium's repository _is_ public, and it's unclear why ".private ==
false" does not evaluate to "false" as we expected in that case. Did the
curl request fail? Did the reply miss the expected .private field? We'll
probably loosen the check as a workaround, but before that it would be
interesting to understand better what's going on. Here we remove the -s
flag from curl and print the reply from the GitHub API request, so we
can better understand what's going on next time we observe a failure.

Signed-off-by: Quentin Monnet <qmo@qmon.net>
Signed-off-by: Jussi Maki <jussi@isovalent.com>